requirements.txt
asgiref==3.8.1
attrs==23.2.0
coverage==7.4.4
Django==5.0.3
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.4.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
